+++ src/App.js
@@ -1,33 +1,15 @@
-import React from 'react';
-
-const todoList = [
-  {
-    id: 1,
-    title: "Completed assignment",
-  },
-  {
-    id: 2,
-    title: "Remove the directory of Completed assignment",
-  },
-  {
-    id: 3,
-    title: "Re-do Coding assignment",
-  },
-];
+import React from "react";
+import TodoList from "./TodoList";
+import AddTodoForm from "./AddTodoForm";
 
 function App() {
   return (
-    <div style={{ textAlign: 'center' }}>
-      <header>
+    <div style={{ textAlign: "center" }}>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (item) {
-          return <li key = {item.id}>{item.title}</li>
-        })}
-      </ul>
-      </header>
+      <AddTodoForm />
+      <TodoList />
     </div>
   );
-}
+};
 
 export default App;
